Iniciando criação de rotinas de auditoria

+++ Aplicacao/phps/produtos_cadastrar2.php
@@ -201,6 +201,15 @@ if ($codigo == "") { //caso seja um cadastro novo fazer isso
         die("Erro22: " . mysql_error());
     $ultimo = mysql_insert_id();
     $produto = $ultimo;
+
+    //Grava Log
+    $sql_logs="
+        INSERT INTO auditoria (aud_usuario_cpf,aud_usuario_nome, aud_operacao, aud_tabela, aud_descricao,aud_sql) 
+        VALUES ('$usuario_cpf','$usuario_nome','INSERT','produtos','Cadastrou um novo produto ($nome) codigo ($ultimo)','$sql')
+    ";
+    if (!$query_logs = mysql_query($sql_logs)) die("Erro ao gravar LOG de auditoria <br>". mysql_error());
+
+
     foreach ($tiponegociacao as $tiponegociacao) {
         $sql2 = "
             INSERT INTO 
@@ -299,6 +308,14 @@ if ($codigo == "") { //caso seja um cadastro novo fazer isso
     ";
     if (!mysql_query($sql))
         die("Erro: " . mysql_error());
+
+    //Grava Log
+    $sql_logs="
+        INSERT INTO auditoria (aud_usuario_cpf,aud_usuario_nome, aud_operacao, aud_tabela, aud_descricao, aud_sql) 
+        VALUES ('$usuario_cpf','$usuario_nome','UPDATE','produtos','Atualizou o produto ($codigo)', '')
+    ";
+    if (!$query_logs = mysql_query($sql_logs)) die("Erro ao gravar LOG de auditoria <br>". mysql_error());
+
     //Deleta os tipos de negociação para depois incluir de novo no novo formato
     $sqldel = " DELETE FROM mestre_produtos_tipo WHERE mesprotip_produto='$codigo'";
     if (!mysql_query($sqldel))
